sprites module created, works
